extract consumer
